+++ src/maicos/modules/velocitycylinder.py
@@ -6,6 +6,7 @@
 """Module for computing cylindrical velocity profiles."""
 
 import logging
+from pathlib import Path
 
 import MDAnalysis as mda
 
@@ -63,7 +64,7 @@ class VelocityCylinder(ProfileCylinderBase):
         pack: bool = True,
         jitter: float = 0.0,
         concfreq: int = 0,
-        output: str = "velocity.dat",
+        output: str | Path = "velocity.dat",
     ) -> None:
         self._locals = locals()
         if vdim not in [0, 1, 2]: